Extract firebase envs
