+++ src/config/firebase.ts
@@ -1,0 +1,14 @@
+import firebase from 'firebase/app';
+import 'firebase/storage';
+import 'firebase/firestore';
+import 'firebase/auth';
+
+import { config } from './index';
+
+firebase.initializeApp(config.firebase);
+
+const auth = firebase.auth();
+const firestore = firebase.firestore();
+const storage = firebase.storage();
+
+export { auth, firestore, storage };
